pyproject.toml
[tool.poetry]
name = "somda_project"
version = "0.1.0"
description = ""
authors = ["Jonas Stettner <mail@jstet.net>"]
readme = "README.md"
packages = [{include = "somda_project"}]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.0.3"
requests = "^2.31.0"
bs4 = "^0.0.1"
modal-client = "^0.53.3660"
minio = "^7.1.15"
python-dotenv = "^1.0.0"
rich = "^13.4.2"
tqdm = "^4.65.0"
pyarrow = "^12.0.1"
duckdb = "^0.8.1"
jupyter = "^1.0.0"
matplotlib = "^3.7.2"
scikit-learn = "^1.3.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
ruff = "^0.0.276"
pre-commit = "^3.3.3"
pytest = "^7.4.0"
importlib-resources = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 120
target-version = ['py310']
preview = true
exclude = '''
/(
    \.git
| \.hg
| \.mypy_cache
| \.tox
| \.venv
| _build
| buck-out
| build
)/
'''

[tool.ruff]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = ["E", "F"]
ignore = []

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["A", "B", "C", "D", "E", "F", "G", "I", "N", "Q", "S", "T", "W", "ANN", "ARG", "BLE", "COM", "DJ", "DTZ", "EM", "ERA", "EXE", "FBT", "ICN", "INP", "ISC", "NPY", "PD", "PGH", "PIE", "PL", "PT", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "TRY", "UP", "YTT"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    "tests/data",
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 120

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"


# Assume Python 3.10.
target-version = "py310"